Move monitor to own file
